[PL-118] refactor: SizedBox to Padding, 위젯명 UpperCase, Listview 수정
